feat: [최현수] 회사별 면접을 시작하고 자기소개 답변을 기반으로 다음 질문을 합니다 [JSAB-104]

+++ src/main/java/com/wowraid/jobspoon/interview/repository/InterviewRepository.java
@@ -3,5 +3,5 @@ package com.wowraid.jobspoon.interview.repository;
 import com.wowraid.jobspoon.interview.entity.Interview;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface InterviewRepository extends JpaRepository<Interview, Integer> {
+public interface InterviewRepository extends JpaRepository<Interview, Long> {
 }
